CLOUDP-196958: Export readiness and versionhook main functions
