+++ pkg/cmd/readiness/testdata/config-current-version.json
