+++ pkg/cmd/readiness/testdata/config-new-version.json
