+++ pkg/cmd/readiness/testdata/health-status-deadlocked-waiting-for-correct-automation-credentials.json
