+++ pkg/cmd/readiness/testdata/health-status-deadlocked-with-prev-config.json
