+++ pkg/cmd/readiness/testdata/health-status-deadlocked.json
