+++ pkg/cmd/readiness/testdata/health-status-enterprise-upgrade-interrupted.json
