+++ pkg/cmd/readiness/testdata/health-status-error-tls.json
