+++ pkg/cmd/readiness/testdata/health-status-no-plans.json
