+++ pkg/cmd/readiness/testdata/health-status-no-replication.json
