+++ pkg/cmd/readiness/testdata/health-status-not-readable-state.json
