+++ pkg/cmd/readiness/testdata/health-status-ok-no-replica-status.json
