+++ pkg/cmd/readiness/testdata/health-status-ok.json
