+++ pkg/cmd/readiness/testdata/health-status-pending.json
