+++ pkg/cmd/readiness/testdata/k8sobjects.go
